Split subscription/mod.rs into logical submodules
Extract the 1000-line monolithic subscription module into 6 focused
submodules while preserving all public API paths and test compatibility:

- core.rs: Subscription trait, BoxedSubscription, Tick/Timer/Channel/Stream
- combinators.rs: Mapped/Filter/Take/Debounce/Throttle subscriptions
- ext.rs: SubscriptionExt trait and blanket impl
- batch.rs: BatchSubscription and batch() function
- interval.rs: IntervalImmediateSubscription and interval_immediate()
- terminal.rs: TerminalEventSubscription and terminal_events()
